kanela support for kamon-annotation
