+++ build.sbt
@@ -1,5 +1,5 @@
 /* =========================================================================================
- * Copyright © 2013-2016 the kamon project <http://kamon.io/>
+ * Copyright © 2013-2018 the kamon project <http://kamon.io/>
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
  * except in compliance with the License. You may obtain a copy of the License at
@@ -13,14 +13,33 @@
  * =========================================================================================
  */
 
-val kamonCore = "io.kamon" %% "kamon-core" % "0.6.7"
-val el = "org.glassfish" % "javax.el" % "3.0.0"
+val kamonCore       = "io.kamon"      %% "kamon-core"     % "1.1.2"
+val kamonTestkit    = "io.kamon"      %% "kamon-testkit"  % "1.1.2"
+val el              = "org.glassfish" %   "javax.el"      % "3.0.0"
 
 lazy val root = (project in file("."))
-  .settings(name := "kamon-annotation")
-  .settings(aspectJSettings: _*)
+  .settings(noPublishing: _*)
+  .aggregate(annotationApi, annotation)
+
+val commonSettings = Seq(
+    scalaVersion := "2.12.6",
+    resolvers += Resolver.mavenLocal,
+    crossScalaVersions := Seq("2.12.6", "2.11.12", "2.10.7")
+)
+
+lazy val annotationApi = (project in file("kamon-annotation-api"))
+  .enablePlugins(JmhPlugin)
+  .settings(moduleName := "kamon-annotation-api", resolvers += Resolver.mavenLocal)
+  .settings(noPublishing: _*)
+  .settings(commonSettings: _*)
+
+lazy val annotation = (project in file("kamon-annotation"))
+  .enablePlugins(JavaAgent)
+  .settings(moduleName := "kamon-annotation")
+  .settings(commonSettings: _*)
+  .settings(javaAgents += "io.kamon"  % "kanela-agent"  % "0.0.300"  % "compile;test")
   .settings(
-    libraryDependencies ++=
-      compileScope(el, kamonCore) ++
-      testScope(scalatest, akkaDependency("testkit").value, slf4jApi, logbackClassic) ++
-      providedScope(aspectJ))
+      libraryDependencies ++=
+        compileScope(kamonCore, el) ++
+          testScope(scalatest, logbackClassic, kamonTestkit)
+  ).dependsOn(annotationApi)
