+++ kamon-annotation/src/main/resources/META-INF/aop.xml
