+++ kamon-annotation/src/main/resources/reference.conf
@@ -1,0 +1,21 @@
+# ========================================= #
+# kamon-annotation Reference Configuration  #
+# ========================================= #
+
+kanela {
+  show-banner = true
+  log-level = "INFO"
+
+  modules {
+    annotation-module {
+      stopable = true
+      name = "Kamon Annotation Module"
+      instrumentations = [
+        "kamon.annotation.instrumentation.AnnotationInstrumentation"
+      ]
+      within = [
+        "kamon.annotation..*",
+      ]
+    }
+  }
+}
