+++ kamon-annotation/src/test/resources/application.conf
