+++ project/plugins.sbt
@@ -1,2 +1,4 @@
 lazy val root: Project = project.in(file(".")).dependsOn(latestSbtUmbrella)
 lazy val latestSbtUmbrella = uri("git://github.com/kamon-io/kamon-sbt-umbrella.git")
+
+addSbtPlugin("com.lightbend.sbt" % "sbt-javaagent" % "0.1.4")
