--- src/main/resources/reference.conf
@@ -1,17 +1,0 @@
-# ========================================= #
-# Kamon-Annotation Reference Configuration  #
-# ========================================= #
-
-kamon {
-  annotation {
-    # We use two arrays to store the kamon instruments in order to do fast array lookups.
-    # These lookups are done using the getId() method on the JoinPoint.StaticPart object.
-    # The ids for all affected join points within a target type are unique (and start from 0).
-    instruments-array-size = 32
-  }
-  modules {
-    kamon-annotation {
-      requires-aspectj = yes
-    }
-  }
-}
