--- src/main/scala/kamon/annotation/AnnotationExtension.scala
@@ -1,27 +1,0 @@
-/* =========================================================================================
- * Copyright © 2013-2015 the kamon project <http://kamon.io/>
- *
- * Licensed under the Apache License, Version 2.0 (the "License") you may not use this file
- * except in compliance with the License. You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the
- * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions
- * and limitations under the License.
- * =========================================================================================
- */
-
-package kamon.annotation
-
-import kamon.Kamon
-import kamon.util.logger.LazyLogger
-
-object AnnotationExtension {
-  val log = LazyLogger("kamon.annotation.AnnotationExtension")
-
-  val config = Kamon.config.getConfig("kamon.annotation")
-  val arraySize = config.getInt("instruments-array-size")
-}
-
